version 4.3.2 and 4.3.3
4.3.3
May, 9th 2008

    To set the default screen through the environment dialog in case of a multiple screens usage wasn't effect, fixed

    Fixe problem in the help dialog.

    When a class stereotyped typedef is read-only and you edit it, the base type is empty in the tab UML and the code preview in C++ tab doesn't show the base type too. Fixed

    The forms @{...} are now substituted in the code generated for the generalizations/realizations.

    Python Generator 1.0.2
    Fixe a bug in the computation of the dependencies to generate the import forms.
    To substitute forms @{...} in the code generated for the generalizations/realizations.

    C++ Generator 2.12.4
    Java Generator 2.15.1
    Php Generator 1.4.1
    Idl Generator 2.9.2
    To substitute forms @{...} in the code generated for the generalizations/realizations.

4.3.2
May, 5th 2008

    Modify the attribution of the socket port number used to exchange data between the modeler and the plug-outs. The goal is to solve problems with firewall sometimes blocking the plug-outs.

    In the previous releases some information was get from environment variables, for instance BOUML_ID. Now these informations are memorized in the file .boumlrc placed into the home directory, and set through a dedicated dialog called through the entry Set environment in the menu Miscellaneous.
    This dialog is automatically called in case .boumlrc doesn't exist, and of course the values set through the defined environment variables are proposed.
    The shortcuts are now saved in the file .bouml_shortcuts

    The chapter of the reference manual corresponding to the type of the element selected in the browser is now shown by using the F1 key or the help menu (About dialog is now associated to the F2 key).
    The chapter starting is used if no project is load, if nothing is selected in the browser the chapter browser items is used.
    The directory where the documentation is placed can be specified by setting the environment (see above), you can also indicate which navigator will be used to show the HTML pages.
    Two setups are now distributed for Windows, one of them install the documentation.

    When you add an inherited operation the new operation is set not abtract. If it is not empty the Php definition is now forced to contain ${body}, to help in case the inherited operation is abstract.

    In the diagrams add go up/down to change a little the level of an element contrarilly to upper and lower.
    Of course you can define shortcuts to these new commands.

    XMI2 Import 1.2.10
    Little improvements.

    Project synchro 1.2
    Project control 1.2
    To not use environment variables. You must use the new release of the modeler to create the file .boumlrc before using these tools.

+++ src/misc/mu.cpp
@@ -30,19 +30,19 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <qdatetime.h>
-#include <qapplication.h>
 
 #include "mu.h"
 #include "myio.h"
 #include "BrowserView.h"
 #include "DialogUtil.h"
+#include "EnvDialog.h"
 
 static int Uid = -1;
 static int RootPermission;	// old Uid or 0
 
 void set_user_id(int i)
 {
-  if (Uid > 0) {
+  if ((Uid > 0) && BrowserView::get_project()) {
     QDir dir = BrowserView::get_dir();
     
     dir.rmdir(QString::number(Uid) + ".lock");
@@ -67,53 +67,21 @@ under the control of 'Project control' or 'Project merge'\n\
     else
       force_read_only(FALSE);
     
-    char * v = getenv("BOUML_ID");
+    Uid = read_boumlrc();
     
-    if ((v == 0) ||
-	(sscanf(v, "%d", &Uid) != 1) ||
-	(Uid < 2) ||
-	(Uid > 127)) {
-      Uid = QTime::currentTime().msec();
-      
-      for (;;) {
-	Uid = Uid % 125 + 2;
-	
-	QString fn = QString::number(Uid) + ".lock";
-	
-	if (dir.mkdir(fn))
-	  break;
-	if (!dir.exists(fn)) {
-	  msg_critical("BOUML_ID", "Can't create directory\n"
-		       + dir.absFilePath(fn) +
-		       ",\nthe project is open in read-only mode");
-	  force_read_only(TRUE);
-	  break;
-	}
-      }
-      
-      msg_critical("BOUML_ID", 
-		   "The BOUML_ID environment variable is not or wrong defined.\n\
-\n\
-This one allows several users to work on the same project, this time your\n\
-identifier is " + QString::number(Uid) +
-		   " but BOUML declines all the responsabilies of this random choice.\n\
-\n\
-You must define the environment variable BOUML_ID valuing between 2 up to 127\n\
-not used by an other person working at the same time on a project with you.");
-    }
-    else {
+    if (BrowserView::get_project()) {
       QString fn = QString::number(Uid) + ".lock";
       
       if (! dir.mkdir(fn)) {
 	if (!dir.exists(fn)) {
-	  msg_critical("BOUML_ID", "Can't create directory "
+	  msg_critical("User Own Identifier", "Can't create directory "
 		       + dir.absFilePath(fn) +
 		       ",\nthe project is open in read-only mode");
-	  force_read_only(TRUE);
+	force_read_only(TRUE);
 	}
 	else {
-	  msg_critical("BOUML_ID", 
-		     "\
+	  msg_critical("User Own Identifier", 
+		       "\
 It seems that you are already editing the project.\n\n\
 If you're SURE that this is not the case and\n\
 another user does not have an identifier equal\n\
